refactor: refactor tab bar component [739]
