+++ src/pages/Collections/Collections.old.svelte
@@ -2,7 +2,7 @@
   import CollectionsList from "$lib/components/collections/collections-list/CollectionList.svelte";
   import RequestResponse from "$lib/components/collections/req-res-section/RequestResponse.svelte";
   import DefaultTabBar from "$lib/components/collections/req-res-section/sub-components/sub-components-header/DefaultTabBar.svelte";
-  import TabBar from "$lib/components/collections/req-res-section/sub-components/sub-components-header/TabBar.svelte";
+  import TabBar from "$lib/components/collections/tab-bar/TabBar.svelte";
   import { collapsibleState } from "$lib/store/request-response-section";
   import type { CollectionsMethods } from "$lib/utils/interfaces/collections.interface";
   import { CollectionsViewModel } from "./Collections.ViewModel.old";
@@ -246,7 +246,7 @@
           {refreshEnv}
         />
       </div>
-      <div class="tab__content d-flex vh-100">
+      <!-- <div class="tab__content d-flex vh-100">
         <div class="w-100">
           {#if $tabList && $tabList.length == 0}
             <DefaultTabBar {collectionsMethods} />
@@ -258,44 +258,44 @@
               {onTabsSwitched}
               {loggedUserRoleInWorkspace}
             />
-          </div>
-          <div class="tab__content d-flex">
-            <div class="w-100">
-              {#if $tabList && $tabList.length == 0}
-                <DefaultTabBar {collectionsMethods} />
-              {:else if $activeTab && $activeTab.type === ItemType.REQUEST}
-                <RequestResponse
-                  {activeTab}
-                  {loggedUserRoleInWorkspace}
-                  {collectionsMethods}
-                  environmentVariables={environmentVariables.reverse()}
-                  {currentWorkspace}
-                />
-              {:else if $activeTab && $activeTab.type === ItemType.WORKSPACE}
-                <MyWorkspace {activeTab} {collectionsMethods} />
-              {:else if $activeTab && $activeTab.type === ItemType.FOLDER}
-                <MyFolder
-                  {collectionsMethods}
-                  {activeTab}
-                  {loggedUserRoleInWorkspace}
-                />
-              {:else if $activeTab && $activeTab.type === ItemType.COLLECTION}
-                <MyCollection
-                  {collectionsMethods}
-                  {activeTab}
-                  {_collectionListViewModel}
-                  {loggedUserRoleInWorkspace}
-                  {currentWorkspaceId}
-                  {currentWorkspace}
-                />
-              {/if}
-            </div>
-          </div>
+          </div> -->
+      <div class="tab__content d-flex">
+        <div class="w-100">
+          {#if $tabList && $tabList.length == 0}
+            <DefaultTabBar {collectionsMethods} />
+          {:else if $activeTab && $activeTab.type === ItemType.REQUEST}
+            <RequestResponse
+              {activeTab}
+              {loggedUserRoleInWorkspace}
+              {collectionsMethods}
+              environmentVariables={environmentVariables.reverse()}
+              {currentWorkspace}
+            />
+          {:else if $activeTab && $activeTab.type === ItemType.WORKSPACE}
+            <MyWorkspace {activeTab} {collectionsMethods} />
+          {:else if $activeTab && $activeTab.type === ItemType.FOLDER}
+            <MyFolder
+              {collectionsMethods}
+              {activeTab}
+              {loggedUserRoleInWorkspace}
+            />
+          {:else if $activeTab && $activeTab.type === ItemType.COLLECTION}
+            <MyCollection
+              {collectionsMethods}
+              {activeTab}
+              {_collectionListViewModel}
+              {loggedUserRoleInWorkspace}
+              {currentWorkspaceId}
+              {currentWorkspace}
+            />
+          {/if}
         </div>
-      </Pane>
-    </Splitpanes>
-  </div>
-</Motion>
+      </div>
+    </div>
+  </Pane>
+</Splitpanes>
+<!-- </div>
+</Motion> -->
 <svelte:window on:keydown={handleKeyPress} />
 
 <style>
